+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -103,6 +103,13 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`bugprone-incorrect-enable-shared-from-this
+  <clang-tidy/checks/bugprone/incorrect-enable-shared-from-this>` check.
+
+  Check if class/structs publicly inherits from ``std::enable_shared_from_this``,
+  because otherwise when ``shared_from_this`` is called unintended behaviour will 
+  occur.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^